Refactor & minor fixes JS
