+++ js/src/product-feed/review-request/index.scss
@@ -29,6 +29,7 @@
 	flex-wrap: wrap;
 	&__text {
 		padding: 0 $grid-unit-20;
+		color: #757575;
 	}
 
 	&__button {
